put LMCacheKVSpec into custom type
