+++ lmcache/integration/tensorrt_llm/tensorrt_mp_adapter.py
@@ -34,9 +34,9 @@ import zmq
 from lmcache import torch_dev
 from lmcache.logging import init_logger
 from lmcache.utils import EngineType, check_interprocess_event_support
-from lmcache.v1.kv_cache_groups import LMCacheKVSpec
 from lmcache.v1.multiprocess.custom_types import (
     IPCCacheEngineKey,
+    LMCacheKVSpec,
     RawCudaIPCWrapper,
 )
 from lmcache.v1.multiprocess.mq import MessageQueueClient, MessagingFuture
