+++ lmcache/v1/multiprocess/transfer_context.py
@@ -13,8 +13,10 @@ from lmcache import torch_dev
 from lmcache.utils import EngineType, init_logger
 from lmcache.v1.distributed.api import MemoryLayoutDesc
 from lmcache.v1.gpu_connector.utils import LayoutHints, is_mla
-from lmcache.v1.kv_cache_groups import LMCacheKVSpec
-from lmcache.v1.multiprocess.custom_types import RegisterNonGpuContextPayload
+from lmcache.v1.multiprocess.custom_types import (
+    LMCacheKVSpec,
+    RegisterNonGpuContextPayload,
+)
 from lmcache.v1.multiprocess.futures import MessagingFuture
 from lmcache.v1.multiprocess.mq import MessageQueueClient
 from lmcache.v1.multiprocess.non_gpu_context import (
