+++ tests/v1/multiprocess/test_mq.py
@@ -13,11 +13,11 @@ import zmq
 
 # First Party
 from lmcache.utils import EngineType
-from lmcache.v1.kv_cache_groups import LMCacheKVSpec
 from lmcache.v1.multiprocess.custom_types import (
     BlockAllocationRecord,
     CudaIPCWrapper,
     IPCCacheEngineKey,
+    LMCacheKVSpec,
 )
 from lmcache.v1.multiprocess.mq import (
     BlockingRequestHandler,
