+++ tests/v1/multiprocess/test_mq_handler_helpers.py
@@ -9,8 +9,11 @@ and passed between processes during multiprocessing tests.
 # First Party
 from lmcache.utils import EngineType
 from lmcache.v1.gpu_connector.utils import LayoutHints
-from lmcache.v1.kv_cache_groups import LMCacheKVSpec
-from lmcache.v1.multiprocess.custom_types import BlockAllocationRecord, KVCache
+from lmcache.v1.multiprocess.custom_types import (
+    BlockAllocationRecord,
+    KVCache,
+    LMCacheKVSpec,
+)
 from lmcache.v1.multiprocess.protocol import KeyType
 
 # ==============================================================================
